requirements.txt
asgiref==3.11.1
Django==5.0.3
django-localflavor==5.0
packaging==26.0
python-stdnum==2.2
sqlparse==0.5.5
tzdata==2025.3
wheel==0.46.3
